+++ lib/@opf_model/init_indexed_name.m
@@ -15,6 +15,7 @@ function om = init_indexed_name(om, set_type, name, dim_list)
 %       SET_TYPE = 'nli'   => nonlinear inequality constraint set
 %       SET_TYPE = 'qdc'   => quadratic cost set
 %       SET_TYPE = 'nlc'   => nonlinear cost set
+%       SET_TYPE = 'cost'  => legacy cost set
 %
 %   Indexed Named Sets
 %
@@ -35,7 +36,7 @@ function om = init_indexed_name(om, set_type, name, dim_list)
 %       end
 %
 %   See also OPT_MODEL, ADD_VAR, ADD_LIN_CONSTRAINT, ADD_NLN_CONSTRAINT,
-%            ADD_QUAD_COST and ADD_NLN_COST.
+%            ADD_QUAD_COST, ADD_NLN_COST and ADD_LEGACY_COST.
 
 %   MATPOWER
 %   Copyright (c) 2008-2020, Power Systems Engineering Research Center (PSERC)
@@ -45,46 +46,18 @@ function om = init_indexed_name(om, set_type, name, dim_list)
 %   Covered by the 3-clause BSD License (see LICENSE file for details).
 %   See https://matpower.org for more info.
 
-%% check for valid type for named set
-st_label = om.valid_named_set_type(set_type);
-if st_label
-    ff = set_type;
-else
-    error('@opt_model/init_indexed_name: ''%s'' is not a valid SET_TYPE, must be one of ''var'', ''lin'', ''nle'', ''nli'', ''cost''', set_type);
-end
-
-%% prevent duplicate name in set of specified type
-if isfield(om.(ff).idx.N, name)
-    error('@opt_model/init_indexed_name: %s set named ''%s'' already exists', ...
-        st_label, name);
-end
-
-%% use column vector if single dimension
-if length(dim_list) == 1
-    dim_list = {dim_list{:}, 1};
-end
-
-%% add general info about this named set
-zero_vector = zeros(dim_list{:});
-empty_cell  = cell(dim_list{:});
-om.(ff).idx.i1.(name)    = zero_vector; %% starting index
-om.(ff).idx.iN.(name)    = zero_vector; %% ending index
-om.(ff).idx.N.(name)     = zero_vector; %% number of vars/constraints/costs
+%% call parent method (also checks for valid type for named set)
+om = init_indexed_name@opt_model(om, set_type, name, dim_list);
 
 %% add type-specific info about this named set
-switch ff
-    case 'var'          %% variable set
-        om.var.data.v0.(name)   = empty_cell;   %% initial value
-        om.var.data.vl.(name)   = empty_cell;   %% lower bound
-        om.var.data.vu.(name)   = empty_cell;   %% upper bound
-        om.var.data.vt.(name)   = empty_cell;   %% variable type
-    case 'lin'          %% linear constraint set
-        om.lin.data.A.(name)   = empty_cell;
-        om.lin.data.l.(name)   = empty_cell;
-        om.lin.data.u.(name)   = empty_cell;
-        om.lin.data.vs.(name)  = empty_cell;
-    case {'nle', 'nli'} %% nonlinear constraint set
-        om.(ff).data.fcn.(name) = empty_cell;
-        om.(ff).data.hess.(name)= empty_cell;
-        om.(ff).data.vs.(name)  = empty_cell;
+switch set_type
+    case 'cost'         %% cost set
+        %% use column vector if single dimension
+        if length(dim_list) == 1
+            dim_list = {dim_list{:}, 1};
+        end
+        empty_cell  = cell(dim_list{:});
+        om.cost.data.N.(name)   = empty_cell;
+        om.cost.data.Cw.(name)  = empty_cell;
+        om.cost.data.vs.(name)  = empty_cell;
 end