+++ lib/t/t_opf_model_legacy.m
@@ -491,7 +491,7 @@ t_ok(isfield(cp, 'N'), t);
 
 %%-----  get_idx  -----
 t = 'get_idx : cost';
-[vv, ll, nne, nni, cc] = om.get_idx();
+cc = om.get_idx('cost');
 t_is([cc.i1.vcost cc.iN.vcost cc.N.vcost], [3 4 2], 14, [t ' : vcost']);
 t_is(size(cc.i1.wc), [2, 2], 14, [t ' : size(cc.i1.wc)']);
 t_is([cc.i1.wc(2,1) cc.iN.wc(2,1) cc.N.wc(2,1)], [10 12 3], 14, [t ' : wc(2,1)']);